# Improve main attack performance