--- a/maps/biter_battles_v2/ai.lua
+++ b/maps/biter_battles_v2/ai.lua
@@ -35,8 +35,6 @@ local threat_values = {
 }
 
 -- these areas are for north
-local left_spawner_area   = {left_top = {-2000, -1200}, right_bottom = {-800, -700}}
-local right_spawner_area  = {left_top = {800,   -1200}, right_bottom = {2000, -700}}
 local middle_spawner_area = {left_top = {-600,  -1000}, right_bottom = {600,  -400}}
 local whole_spawner_area  = {left_top = {-2048, -1400}, right_bottom = {2048, -400}}