This a rebased version (with two commits squashed) of #116.
I have tested it on map at 86% evo from Sunday and let it play out normally. Everything looks fine.
```
## 3600 ticks; 1 attack cycle
# pre
avg: 8.379 ms, min: 6.126 ms, max: 537.199 ms
avg: 8.533 ms, min: 6.064 ms, max: 538.407 ms
# post
avg: 8.178 ms, min: 6.050 ms, max: 106.591 ms
avg: 8.285 ms, min: 6.077 ms, max: 98.863 ms

## 7200 ticks; 2 attack cycles
# pre
avg: 8.192 ms, min: 5.940 ms, max: 543.868 ms
avg: 8.319 ms, min: 5.938 ms, max: 543.673 ms
# post
avg: 8.014 ms, min: 5.691 ms, max: 109.667 ms
avg: 8.125 ms, min: 5.953 ms, max: 107.933 ms
```